requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.4.0
Pillow==10.2.0
pydantic==2.6.1
httpx==0.27.0
groq
